Tkcrm 20 raw data student

+++ app/Http/Controllers/ClientStudentController.php
@@ -791,7 +791,14 @@ class ClientStudentController extends ClientController
         try {
 
             $rawClient = $this->clientRepository->getViewRawClientById($rawClientId);
-            $clientId != null ? $client = $this->clientRepository->getViewClientById($clientId) : null;
+            if (!isset($rawClient))
+                return Redirect::to('client/student/raw')->withError('Data does not exist');
+
+            if ($clientId != null){
+                $client = $this->clientRepository->getViewClientById($clientId);
+                if (!isset($client))
+                    return Redirect::to('client/student/raw')->withError('Data does not exist');
+            }
 
             DB::commit();
         } catch (Exception $e) {
@@ -942,6 +949,9 @@ class ClientStudentController extends ClientController
 
         DB::beginTransaction();
         try {
+
+            if (!isset($rawStudent))
+                return Redirect::to('client/student/raw')->withError('Data does not exist');
 
             $this->clientRepository->deleteClient($rawclientId);
             DB::commit();
